Create DisableSelfPurchaseOffice365
This script disables all products for self-purchase option.  We start by loading the necessary modules, then retrieve all products in the self service policy then loop through that array setting each to false.  The out-string line converts the system.object variable type to system.text as required in the command to change the status fixing the fault in the original script from Microsoft